# api/openapi.yaml
openapi: 3.1.0
info:
  title: Wealthboard API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
paths:
...
  /ai/settings:
    put:
      operationId: saveAISettings
      summary: Save the current user's AI provider settings
      description: Requires a browser session, trusted Origin, and `X-CSRF-Token`; Bearer API keys are not accepted. The JSON body is limited to 16 KiB. Stored API key ciphertext is never returned.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/MutationCSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AISettingsInput" }
      responses:
        "200":
          description: Saved settings with redacted credential metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AIWorkflowSettings" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        "504": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: disconnectAIProvider
      summary: Delete AI settings and disconnect the provider
      description: "Requires a browser session, trusted Origin, and `X-CSRF-Token`; Bearer API keys are not accepted. Removes owner-scoped settings and stored credential material and returns `Cache-Control: no-store`."
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/MutationCSRFToken" }]
      responses:
        "204":
          description: Provider disconnected
          headers: { Cache-Control: { $ref: "#/components/headers/NoStore" } }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        "504": { $ref: "#/components/responses/Problem" }
  /ai/credential:
    post:
      operationId: saveAICredential
      summary: Encrypt and store an AI provider API key
      description: Requires a browser session, trusted Origin, and `X-CSRF-Token`; Bearer API keys are not accepted. The JSON body is limited to 8 KiB. The plaintext key is used only for encryption and is never returned or logged.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/MutationCSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AICredentialInput" }
      responses:
        "200":
          description: Settings with redacted credential metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AIWorkflowSettings" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        "504": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteAICredential
      summary: Delete the stored AI provider credential
      description: "Requires a browser session, trusted Origin, and `X-CSRF-Token`; Bearer API keys are not accepted. Deletes only the current user's encrypted credential and returns `Cache-Control: no-store`."
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/MutationCSRFToken" }]
      responses:
        "204":
          description: Stored credential deleted
          headers: { Cache-Control: { $ref: "#/components/headers/NoStore" } }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        "504": { $ref: "#/components/responses/Problem" }
  /ai/usage/clear:
    post:
      operationId: clearAIUsage
      summary: Delete the current user's AI usage history
      description: Requires a browser session, trusted Origin, and `X-CSRF-Token`; Bearer API keys are not accepted. Only owner-scoped usage events are deleted.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/MutationCSRFToken" }]
      responses:
        "200":
          description: Number of usage events deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AIClearUsageResult" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        "504": { $ref: "#/components/responses/Problem" }
  /ai/review:
    post:
      operationId: createAIReview
      summary: Generate a review from a supplied portfolio snapshot
      description: Bearer API keys require `ai:invoke`. Browser sessions require a trusted Origin and `X-CSRF-Token`. The JSON body is limited to 40 KiB and the embedded snapshot to 25,000 bytes. Exact amounts and account names are sent only when the explicit sharing flags permit them. A request API key is ephemeral and never returned, stored, or logged.
      security: [{ sessionCookie: [] }, { apiKey: [] }]
      parameters: [{ $ref: "#/components/parameters/MutationCSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AIReviewRequest" }
      responses:
        "200":
          description: Grounded review, echoed snapshot, redacted provider metadata, and usage
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AIReviewResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        "504": { $ref: "#/components/responses/Problem" }
  /ai/import/extract:
    post:
      operationId: extractAIImportSource
      summary: Extract bounded text units from one import source file
      description: Bearer API keys require `ai:invoke`. Browser sessions require a trusted Origin and `X-CSRF-Token`. Accepts exactly one file up to 5 MiB plus at most one PDF password; multipart overhead is limited to 64 KiB. Supported files are TXT, CSV, TSV, JSON, PDF, XLSX, and DOCX. Extracted text is limited to 64 KiB and 1,000 units. Passwords are cleared after use and never returned, stored, or logged.
      security: [{ sessionCookie: [] }, { apiKey: [] }]
      parameters: [{ $ref: "#/components/parameters/MutationCSRFToken" }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/AISourceUpload" }
      responses:
        "200":
          description: Extracted, bounded source units
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AISourceResult" }
        "400":
          description: Invalid multipart source or a structured extraction error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
            application/json:
              schema: { $ref: "#/components/schemas/AISourceError" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        "504": { $ref: "#/components/responses/Problem" }
  /ai/import/convert:
    post:
      operationId: convertAIImportSource
      summary: Convert extracted source units into an import draft
      description: Bearer API keys require `ai:invoke`. Browser sessions require a trusted Origin and `X-CSRF-Token`. The JSON body is limited to 5 MiB plus 128 KiB, while the encoded source itself remains limited to 64 KiB and 1,000 units. Consent and a matching settings/destination hash are mandatory. Source text is treated as untrusted data. An ephemeral API key is never returned, stored, or logged.
      security: [{ sessionCookie: [] }, { apiKey: [] }]
      parameters: [{ $ref: "#/components/parameters/MutationCSRFToken" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AIConversionRequest" }
      responses:
        "200":
          description: Import draft and complete source-reference accounting
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AIConversionDraft" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "502": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
        "504": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: wealthboard_session
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: wbk_v1
      description: API-key scopes are enforced per operation as documented in each operation description.
  parameters:
...
    MutationCSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required when authenticating with `sessionCookie`; ignored for Bearer API keys.
      schema:
        type: string
        minLength: 1
...
  headers:
    NoStore:
      description: Prevents storage of private financial or credential metadata.
      schema:
        type: string
        const: no-store
...
  schemas:
...
    CurrencyCode:
      type: string
      pattern: "^[A-Z]{3}$"
...
    AIProvider:
      type: string
      enum: [openai, deepseek, custom]
    AISettingsInput:
      type: object
      additionalProperties: false
      required:
        [
          provider,
          baseUrl,
          model,
          includeExactAmounts,
          includeAccountNames,
          monthlyTokenLimit,
          maxOutputTokens,
        ]
      properties:
        provider: { $ref: "#/components/schemas/AIProvider" }
        baseUrl: { type: string, format: uri }
        model: { type: string, minLength: 1, maxLength: 200 }
        includeExactAmounts:
          {
            type: boolean,
            description: Whether exact financial amounts may be shared with the configured provider.,
          }
        includeAccountNames:
          {
            type: boolean,
            description: Whether private account names may be shared with the configured provider.,
          }
        monthlyTokenLimit: { type: integer, minimum: 10000, maximum: 100000000 }
        maxOutputTokens:
          {
            type: integer,
            minimum: 256,
            description: Must not exceed `monthlyTokenLimit`.,
          }
    AIWorkflowSettings:
      type: object
      additionalProperties: false
      required:
        [
          provider,
          baseUrl,
          model,
          includeExactAmounts,
          includeAccountNames,
          monthlyTokenLimit,
          maxOutputTokens,
          hasStoredApiKey,
          apiKeyHint,
          updatedAt,
        ]
      properties:
        provider: { $ref: "#/components/schemas/AIProvider" }
        baseUrl: { type: string, format: uri }
        model: { type: string }
        includeExactAmounts: { type: boolean }
        includeAccountNames: { type: boolean }
        monthlyTokenLimit: { type: integer }
        maxOutputTokens: { type: integer }
        hasStoredApiKey: { type: boolean }
        apiKeyHint:
          {
            type: [string, "null"],
            description: Non-secret credential hint; plaintext and ciphertext are never returned.,
          }
        updatedAt: { type: string, format: date-time }
    AICredentialInput:
      type: object
      additionalProperties: false
      required: [apiKey]
      properties:
        apiKey: { type: string, minLength: 1, maxLength: 4096, writeOnly: true }
    AIClearUsageResult:
      type: object
      additionalProperties: false
      required: [deleted]
      properties:
        deleted: { type: integer, format: int64, minimum: 0 }
    AIReviewPeriod:
      type: string
      enum: [1m, 3m, 6m, 1y, all]
    AIReviewFocus:
      type: string
      enum: [overall, allocation, goals, cash-flow, data-quality]
    AIReviewSnapshot:
      type: object
      additionalProperties: false
      required:
        [
          schemaVersion,
          asOf,
          period,
          focus,
          baseCurrency,
          sharing,
          completeness,
          portfolio,
          allocations,
          topAccounts,
          cashFlow,
          goals,
          dataQuality,
          methodology,
        ]
      properties:
        schemaVersion: { type: integer, const: 1 }
        asOf: { type: string, minLength: 1 }
        period: { $ref: "#/components/schemas/AIReviewPeriod" }
        focus: { $ref: "#/components/schemas/AIReviewFocus" }
        baseCurrency: { $ref: "#/components/schemas/CurrencyCode" }
        sharing: { type: object, additionalProperties: true }
        completeness: { type: object, additionalProperties: true }
        portfolio: { type: object, additionalProperties: true }
        allocations: { type: [object, array] }
        topAccounts: { type: array, items: {} }
        cashFlow: { type: object, additionalProperties: true }
        goals: { type: array, items: {} }
        dataQuality: { type: [object, array] }
        methodology: { type: object, additionalProperties: true }
    AIReviewRequest:
      type: object
      additionalProperties: false
      required:
        [period, focus, includeExactAmounts, includeAccountNames, snapshot]
      properties:
        period: { $ref: "#/components/schemas/AIReviewPeriod" }
        focus: { $ref: "#/components/schemas/AIReviewFocus" }
        includeExactAmounts: { type: boolean }
        includeAccountNames: { type: boolean }
        apiKey:
          {
            type: string,
            minLength: 8,
            maxLength: 4096,
            writeOnly: true,
            description: Ephemeral provider key; used only when no stored credential is selected.,
          }
        snapshot: { $ref: "#/components/schemas/AIReviewSnapshot" }
    AIReviewFinding:
      type: object
      additionalProperties: false
      required:
        [id, category, severity, confidence, title, explanation, evidenceRefs]
      properties:
        id: { type: string, minLength: 1, maxLength: 80 }
        category:
          {
            type: string,
            enum:
              [data-quality, allocation, liquidity, cash-flow, goals, general],
          }
        severity: { type: string, enum: [info, attention, high] }
        confidence: { type: string, enum: [low, medium, high] }
        title: { type: string, minLength: 1, maxLength: 140 }
        explanation: { type: string, minLength: 1, maxLength: 700 }
        evidenceRefs:
          { type: array, minItems: 1, maxItems: 6, items: { type: string } }
    AIReview:
      type: object
      additionalProperties: false
      required:
        [
          schemaVersion,
          headline,
          executiveSummary,
          dataQuality,
          strengths,
          attentionItems,
          goalObservations,
          questions,
          possibleNextChecks,
          limitations,
        ]
      properties:
        schemaVersion: { type: integer, const: 1 }
        headline: { type: string, minLength: 1, maxLength: 160 }
        executiveSummary: { type: string, minLength: 1, maxLength: 1400 }
        dataQuality:
          {
            type: array,
            maxItems: 6,
            items: { $ref: "#/components/schemas/AIReviewFinding" },
          }
        strengths:
          {
            type: array,
            maxItems: 6,
            items: { $ref: "#/components/schemas/AIReviewFinding" },
          }
        attentionItems:
          {
            type: array,
            maxItems: 8,
            items: { $ref: "#/components/schemas/AIReviewFinding" },
          }
        goalObservations:
          {
            type: array,
            maxItems: 6,
            items: { $ref: "#/components/schemas/AIReviewFinding" },
          }
        questions:
          { type: array, maxItems: 6, items: { type: string, maxLength: 300 } }
        possibleNextChecks:
          { type: array, maxItems: 6, items: { type: string, maxLength: 300 } }
        limitations:
          {
            type: array,
            minItems: 1,
            maxItems: 8,
            items: { type: string, minLength: 1, maxLength: 300 },
          }
    AIReviewResult:
      type: object
      additionalProperties: false
      required: [review, snapshot, provider, usage, generatedAt]
      properties:
        review: { $ref: "#/components/schemas/AIReview" }
        snapshot: { $ref: "#/components/schemas/AIReviewSnapshot" }
        provider:
          type: object
          additionalProperties: false
          required: [name, host, model]
          properties:
            name: { $ref: "#/components/schemas/AIProvider" }
            host: { type: string }
            model: { type: string }
        usage:
          type: object
          additionalProperties: false
          required: [inputTokens, outputTokens]
          properties:
            inputTokens: { type: [integer, "null"] }
            outputTokens: { type: [integer, "null"] }
        generatedAt: { type: string, format: date-time }
    AISourceUpload:
      type: object
      additionalProperties: false
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: One TXT, CSV, TSV, JSON, PDF, XLSX, or DOCX file, maximum 5 MiB.
        documentPassword:
          {
            type: string,
            maxLength: 1024,
            writeOnly: true,
            description: Optional and accepted for PDF files only.,
          }
    AISourceUnit:
      type: object
      additionalProperties: false
      required: [id, location, text]
      properties:
        id: { type: string, pattern: "^source-[0-9]+$" }
        location: { type: string, minLength: 1, maxLength: 200 }
        text: { type: string, minLength: 1, maxLength: 65536 }
    AISource:
      type: object
      additionalProperties: false
      required: [units, warnings]
      properties:
        units:
          {
            type: array,
            minItems: 1,
            maxItems: 1000,
            items: { $ref: "#/components/schemas/AISourceUnit" },
          }
        warnings: { type: array, maxItems: 100, items: { type: string } }
    AISourceResult:
      type: object
      additionalProperties: false
      required: [source]
      properties:
        source: { $ref: "#/components/schemas/AISource" }
    AISourceError:
      type: object
      additionalProperties: false
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: string }
    AIConversionRequest:
      type: object
      additionalProperties: false
      required: [source, configurationHash, consent, trackingMode, currency]
      properties:
        source: { $ref: "#/components/schemas/AISource" }
        configurationHash: { type: string, pattern: "^[a-f0-9]{64}$" }
        consent: { type: boolean, const: true }
        apiKey: { type: string, minLength: 8, maxLength: 4096, writeOnly: true }
        trackingMode: { type: string, enum: [balance, positions] }
        currency: { $ref: "#/components/schemas/CurrencyCode" }
    AIDraftReference:
      type: object
      additionalProperties: false
      required: [collection, row, sourceIds]
      properties:
        collection: { type: string }
        row: { type: integer, minimum: 1 }
        sourceIds:
          {
            type: array,
            minItems: 1,
            items: { type: string, pattern: "^source-[0-9]+$" },
          }
    AIExclusion:
      type: object
      additionalProperties: false
      required: [sourceId, reason]
      properties:
        sourceId: { type: string, pattern: "^source-[0-9]+$" }
        reason: { type: string }
    AIConversionDraft:
      type: object
      additionalProperties: false
      required: [content, references, exclusions, issues]
      properties:
        content:
          {
            type: string,
            description: Generated Wealthboard import document content.,
          }
        references:
          {
            type: array,
            items: { $ref: "#/components/schemas/AIDraftReference" },
          }
        exclusions:
          { type: array, items: { $ref: "#/components/schemas/AIExclusion" } }
        issues: { type: array, maxItems: 100, items: { type: string } }
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
  responses:
...
    Problem:
      description: Request failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"